Ascending search for rulesets feature added
Looking for ruleset modules at first in project packages and then in the add
repository function. Searching happens in an ascending manner where a
ruleset with the filename "comdaan_ruleset.py" is searched for in the
repository itself and every parent of it. An error message is displayed
if the ruleset file is erroneous without stopping the execution.