Multiprocessing the work on entries in a repository
Adopting an aproach where the analysis itself of each entry of a repository
happens in parallel.
Using apply or apply_async yielded longer execution times with
multiprocessing. So, switching to a wrapper function (wrapper) that takes
a path as its only argument and mapping it to self.__paths. This makes
the execution time smaller than without multiprocessing and allows near
maximum cpu usage.
Modules not found errors due to multiprocessing fixed with adding
directory paths to sys.path so even after forking, the system still
recognizes the modules.
Pickling issue encountered fixed with using pathos' multiprocess instead
of the built-in multiprocessing.

Requirements added to pipfile.

As for pool parameters:
After testing different execution times, a pool with 6 (cpu_count + 2)
processes turned out to be the best option. In fact, since we're mainly
IO bound, processes spend most of their times in a blocked state and a
process in a blockd state lets others run, so having more processes
running than cores makes sense in this case.

Removal of the get_log_from_repositories function to remain faithful to
the original API.