Script for fetching git issues
Using python-gitlab to fetch the issues of a project with a given
project id.
Adding a discussion field to the issue dictionary that stores the
discussion around said issue. And thus it stores a list or threads,
which are a list of comments... so a list of lists of dictionaries.
Adding private access token as mandatory cli argument for gitlab
instance authentication.
Shifting to Pipfile, Requirements.txt removed