Network analysis for mailing lists
Heavily inspired from the git repositories' netwrok analysis, it
expresses how much a sender's message was referenced and how much they
reference other messages.
Instead of the intersection of the references lists of each user like in
the git version, this version's edge list weights is the
intersection of the list of message ids of all of the source's message
and the references list of the target. It thus focuses on the sender
being referenced instead of just the sender referencing.
And thus we consider the phenomenon of being referenced since the former is
be more meaningful as to the sender/person's value inside the mailing
community should increase with the amount of time they are referenced.

Moreover in the process of sending and receiving emails some people are
mentioned and replied to and some others mention and reply, and since
being mentioned many times is a more important indicator to someone's
centrality than barely mentioning (because, really, anyone can mention
others), I'm switching to a firected graph to represent the data.