Network Analysis for issue discussions
The network analysis, when it comes to issues is based on the
discussions that happen on the issue themselves.
And so, we build a graph that connects the author of an issue to all
those that commented on said issue.
Not removing edges with a weight of 0 because simply, there
aren't any, at least not with the way the edge_list is currently built.