Introducing LOWESS regression for teamsize analysis
Using the LOWESS regression to calculate a smooth regression for both
commit count and team size. The data fraction used for each estimation
is a function of the length of the abcissa list. And there is no
residual based reweighting done since the noise isn't gravely heavy
tailed.
The results are very similar to those of the rolling window regression.

LOWESS plotting using "date" for the x axis instead of the returned x
axis values by the lowess regression function since the values the x axis,
after conversion from timestamp, are exactly the same as the dates in the
"dates" column in teamsize. So, there is no need for the *_lowess_x
column and the conversion of the returned lowess x axis values.