LOWESS integration for mail centrality analysis
As to for lists that span over less than the centrality step handled:
Mailing lists that span over less than twice the relativedelta end up
having their end date inferir to their start date. The previous fix
which consisted of not adding nor substracting said delta to the min and
max date before calculating the date range made it so that the start
date and the end date in the dates list were respectively lower and
higher than the actual ones.
To fix this issue, for such small mailing lists the relativedelta is
smaller (a week) and the frequency is set to weekly.

Additionally, calculating dates for a centrality analysis in such a way
doesn't take into account the entry corresponding to the maximum date
when it falls slightly after a week/month after the date_range's
max_date (max - relativedelta) since the last date of the dates list
isn't max_date + relativedelta but the date before it in date_range +
relativedelta.

Adding a date interval consisting of dates's last date and emails["date"] to
actually cover the whole date range in emails.