Display function added for use with all types of analyses
One display function for all types of analyses.
In this version different types of analyses cannot be displayed together
because of the different types of display on one hand (graph vs list vs
plot) and on the other because of the confusion that might ensue
regarding the axes.

_display function added to display module because it doesn't really
belong in all of its detailedness in comdaan.py.
display function now displaying different analyses types on the same
page in a gridplot.

object_types module added to avoid circular dependencies in imports.

Minor adjustmnts to the functions relating to the analyses so display
can work with the resutling object types and to the correspondng tests as
well.

Issues JSON file added for testing purposes.