Extracting processing specific to issues from activity and teamsize
Extracting processing specific to issues from activity and teamsize to
an external function.
Now the activity and teamsize functions don't have a "rep" and "com"
condition to deal with reporters and commenters. Instead the user, if
they want the commenter data analyzed has to parse the comments with the
parse_comments function.

Corresponding changes to tests.